refactor: 현재 사용자 상태 users -> currentUsers 이름 변경, axios Users타입명 수정
- 현재 사용자 상태를 users에서 currentUsers로 이름 변경

- axios Users의 타입을 GetUsersResponse로 변경
