+++ src/stores/currentUser.ts
